// module31_homework/my-nest-app/src/modules/addCar/addcar.interface.ts
export interface AddCar {
    brand: String;
    model: String;
    yearIssue: Number;
    governmentNumber: String;
    vin: String;
    color: String;
    drive: String;
    Enginefuel: String;
    Enginevolume: String;
    Enginepower: String;
    transmission: String;
    mileage: String;
    VPSerialAndNumber: String;
    VRCSerialAndNumber: String;
    PriceOne: String;
    PriceTwo: String;
    PriceThree: String;
    Osago: String;
    Kasko: String;
    Options: [];
    AdditionalServices: [];
    photosAuto: [];
    photosDoc: [];
}